Fix the issue that Kafka v2 source may return duplicated records when  is
